+++ FVNever.Reuse/Commenters/ICommenter.cs
@@ -10,10 +10,10 @@ public interface ICommenter
     /// <summary>
     /// Generate a string with metadata information that will be inserted at the beginning of the file.
     /// </summary>
-    /// <param name="copyrightStatements">The copyright statement collection.</param>
+    /// <param name="copyrightNotices">The copyright notice collection.</param>
     /// <param name="licenseIdentifiers">The license identifier collection.</param>
     /// <returns>Resulting string, ending with a newline.</returns>
-    string GenerateHeader(IEnumerable<CopyrightStatement> copyrightStatements, IEnumerable<string> licenseIdentifiers);
+    string GenerateHeader(IEnumerable<CopyrightNotice> copyrightNotices, IEnumerable<string> licenseIdentifiers);
 
     /// <summary>Remove metadata information from the file contents.</summary>
     /// <param name="currentContent">Current file contents.</param>