+++ FVNever.Reuse/ReuseDirectory.cs
@@ -5,6 +5,7 @@
 using System.Text;
 using FVNever.Reuse.Dep5;
 using GitignoreParserNet;
+using JetBrains.Annotations;
 using Microsoft.Extensions.FileSystemGlobbing;
 using TruePath;
 
@@ -14,6 +15,7 @@ namespace FVNever.Reuse;
 /// Provides high-level operations for scanning a directory and extracting REUSE licensing information
 /// from files, sidecar <c>.license</c> files, and DEP-5 metadata.
 /// </summary>
+[PublicAPI]
 public static class ReuseDirectory
 {
     /// <summary>
@@ -24,6 +26,12 @@ public static class ReuseDirectory
     /// A task that produces a list of <see cref="ReuseFileEntry"/> values. Each entry corresponds to a file
     /// for which licensing information was found in-place, in a sidecar <c>.license</c> file, or via a matching DEP-5 stanza.
     /// </returns>
+    /// <remarks>
+    /// Note that for related formats that don't follow the REUSE specification strictly, e.g., the
+    /// <a href="https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/">DEP5 machine-readable
+    /// <c>debian/copyright</c> file contents</a>, this method does its best to parse the copyright notices in the
+    /// typical form they are provided, but can't guarantee their correctness.
+    /// </remarks>
     public static async Task<List<ReuseFileEntry>> ReadEntries(AbsolutePath directory)
     {
         var allFiles = await EnumerateFiles(directory).ConfigureAwait(false);