+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased] (2.0.0)
 ### Changed
-- Copyright statements are now represented by `CopyrightStatement` class instead of raw strings.
+- Copyright notices are now represented by `CopyrightNotice` class instead of raw strings.
 
 ## [1.1.0] - 2025-08-24
 ### Added